+++ src/components/utils.ts
@@ -781,6 +781,25 @@ export const isIosDevice =
     (window.navigator.platform === 'MacIntel' && window.navigator.maxTouchPoints > 1));
 
 /**
+ * Compares two values with unknown type
+ *
+ * @param var1 - value to compare
+ * @param var2 - value to compare with
+ * @returns {boolean} true if they are equal
+ */
+export function equals(var1: unknown, var2: unknown): boolean {
+  const isVar1NonPrimitive = Array.isArray(var1) || isObject(var1);
+  const isVar2NonPrimitive = Array.isArray(var2) || isObject(var2);
+
+  if (isVar1NonPrimitive || isVar2NonPrimitive) {
+    return JSON.stringify(var1) === JSON.stringify(var2);
+  }
+
+  return var1 === var2;
+}
+
+
+/**
  * Returns production of multiplication of arguments
  *
  * @param numbers - numbers to multiply