+++ types/index.d.ts
@@ -48,6 +48,7 @@ export {
   Tool,
   ToolConstructable,
   ToolboxConfig,
+  ToolboxConfigEntry,
   ToolSettings,
   ToolConfig,
   PasteEvent,
@@ -70,6 +71,9 @@ export {
   Dictionary,
   DictValue,
   I18nConfig,
+  PopoverItem,
+  PopoverItemWithConfirmation,
+  PopoverItemWithoutConfirmation
 } from './configs';
 export {OutputData, OutputBlockData} from './data-formats/output-data';
 export { BlockAPI } from './api'